Removed Tabs from downloads activity.
