--- app/src/main/java/in/co/rajkumaar/amritarepo/downloads/adapters/CategoryAdapter.java
@@ -1,75 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2018  RAJKUMAR S
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package in.co.rajkumaar.amritarepo.downloads.adapters;
-
-import android.content.Context;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentPagerAdapter;
-
-import in.co.rajkumaar.amritarepo.R;
-import in.co.rajkumaar.amritarepo.downloads.fragments.DocumentsFragment;
-import in.co.rajkumaar.amritarepo.downloads.fragments.ImagesFragment;
-
-public class CategoryAdapter extends FragmentPagerAdapter {
-    private Context mContext;
-
-    public CategoryAdapter(Context context, FragmentManager fm) {
-        super(fm);
-        mContext = context;
-    }
-
-    /**
-     * Return the {@link Fragment} that should be displayed for the given page number.
-     */
-    @Override
-    public Fragment getItem(int position) {
-        if (position == 0) {
-            return new DocumentsFragment();
-        } else {
-            return new ImagesFragment();
-
-        }
-    }
-
-    /**
-     * Return the total number of pages.
-     */
-    @Override
-    public int getCount() {
-        return 2;
-    }
-
-    @Nullable
-    @Override
-    public CharSequence getPageTitle(int position) {
-        if (position == 0) {
-            return mContext.getString(R.string.category_docs);
-        } else {
-            return mContext.getString(R.string.category_images);
-        }
-    }
-}
